% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{get_unique_choices}
\alias{get_unique_choices}
\title{Get unique choices from a vector input}
\usage{
get_unique_choices(vec)
}
\arguments{
\item{vec}{a vector of characters}
}
\value{
sorted unique vector with no empty strings
}
\description{
Get unique choices from a vector input
}
