% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load_and_prepare_data}
\alias{load_and_prepare_data}
\title{function that wraps loading and parsing to clean the server script}
\usage{
load_and_prepare_data(file)
}
\arguments{
\item{file}{a file path or shiny input list}
}
\value{
prepared data.frame
}
\description{
function that wraps loading and parsing to clean the server script
}
