% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{parse_uploaded_file}
\alias{parse_uploaded_file}
\title{Parse uploaded Excel file}
\usage{
parse_uploaded_file(file)
}
\arguments{
\item{file}{An input file for shiny}
}
\value{
A data.frame
}
\description{
Parse uploaded Excel file
}
