% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{prepare_shiny_data}
\alias{prepare_shiny_data}
\title{Prepare Data for Shiny app}
\usage{
prepare_shiny_data(df)
}
\arguments{
\item{df}{a data.frame, the same one previously parsed}
}
\value{
a data.frame with the any existing columns intact and two added columns
}
\description{
Takes the full data.frame and then subsets the 3 static columns and adds empty
columns for the new information.
}
