% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{update_choices}
\alias{update_choices}
\title{Update reactive choices in state from a data frame}
\usage{
update_choices(state, df)
}
\arguments{
\item{state}{reactiveValues object containing label_choices and category_choices}

\item{df}{data.frame with Standared_Node_names and categories columns}
}
\description{
Update reactive choices in state from a data frame
}
